% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{bfrr}
\alias{bfrr}
\title{Title}
\usage{
bfrr(
  data_model,
  alternativeprior,
  nullprior,
  parameters,
  precision,
  cutoff,
  multicore
)
}
\arguments{
\item{multicore}{}
}
\value{

}
\description{
Title
}
