% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_functions.R
\name{extract_predictions}
\alias{extract_predictions}
\title{Extract predictions}
\usage{
extract_predictions(x)
}
\arguments{
\item{x}{a \code{product} object}
}
\value{
a \code{prediction} object
}
\description{
Extract the marginal predictions over the prior
}
