% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{integral}
\alias{integral}
\title{Compute integral}
\usage{
integral(obj)
}
\arguments{
\item{obj}{a \code{predictive} object}
}
\value{
A numeric of the point predictive value
}
\description{
Computes the definite integral of a \code{predictive} object over the range
of the parameter
}
