% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{likelihood}
\alias{likelihood}
\title{Specify a likelihood}
\usage{
likelihood(distribution, ...)
}
\arguments{
\item{distribution}{the likelihood distribution (see details)}

\item{...}{see details}
}
\value{
an object of class \code{likelihood}
}
\description{
Specify a likelihood
}
\details{
\subsection{Available distributions}{

The following distributions can be used for the likelihood
\itemize{
\item \code{normal} a normal distribution
\item \code{student_t} a scaled and shifted t-distribution
\item \code{noncentral_t} a noncentral t distribution (t scaled)
\item \code{noncentral_d} a noncentral t distribution (d scaled)
The parameters that need to be specified will be dependent on the
distribution
}
}

\subsection{normal distribution}{

When \code{distribution} is set to \code{normal} then the following
parameters must be set
\itemize{
\item \code{mean} mean of the normal likelihood
\item \code{sd} standard deviation of the normal likelihood
}
}

\subsection{student_t distribution}{

When \code{distribution} is set to \code{student_t} then the following
parameters may be set
\itemize{
\item \code{mean} mean of the scaled and shifted t likelihood
\item \code{sd} standard deviation of the scaled and shifted t likelihood
\item \code{df} degrees of freedom
}
}

\subsection{noncentral_t distribution}{

When \code{distribution} is set to \code{noncentral_t} then the following
parameters may be set
\itemize{
\item \code{t} the t value of the data
\item \code{df} degrees of freedom
}
}

\subsection{noncentral_d distribution}{

When \code{distribution} is set to \code{noncentral_d} then the following
parameters may be set
\itemize{
\item \code{d} the d (mean / sd) value of the data
\item \code{df} degrees of freedom
}
}
}
\examples{
# specify a normal likelihood
likelihood(distribution = "normal", mean = 5.5, sd = 32.35)

# specify a scaled and shifted t likelihood
likelihood(distribution = "student_t", mean = 5.5, sd = 32.35, df = 10)

# specify non-central t likelihood (t scaled)
likelihood(distribution = "noncentral_t", t = 10, df = 10)

# specify non-central t likelihood (d scaled)
likelihood(distribution = "noncentral_d", d = 10, df = 10)
}
