% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{names,bayesplay-method}
\alias{names,bayesplay-method}
\title{Get names from data slot}
\usage{
\S4method{names}{bayesplay}(x)
}
\arguments{
\item{x}{a \code{bayesplay} object}
}
\value{
the field names from the data slot
}
\description{
Get names from data slot
}
