% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot.prior}
\alias{plot.posterior}
\title{somethjing}
\usage{
plot(x, ...)

\method{plot}{prior}(x)

\method{plot}{posterior}(x, add_prior = FALSE)
}
\arguments{
\item{x}{a \code{likelihood}, \code{prior}, \code{posterior}, \code{product} or \code{predictive} object}
}
\description{
Plots an object created by bayesplay
}
