% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{posterior_samples}
\alias{posterior_samples}
\title{Draw samples from a posterior distribution}
\usage{
posterior_samples(posterior_obj, n, lower, upper)
}
\arguments{
\item{posterior_obj}{a \code{posterior} object}

\item{n}{the number of samples to draw}

\item{lower}{the lower bound of the posterior distribution}

\item{upper}{the upper bound of the posterior distribution}
}
\value{
A vector of samples from the posterior distribution
}
\description{
Draw samples from a \code{posterior} object using
rejection sampling
}
\examples{
# define a likelihood
data_model <- likelihood(family = "binomial", 10, 20)

# define a prior
prior_model <- prior(family = "beta", 10, 10)

# compute a posterior
post <- extract_posterior(data_model * prior_model)

# Draw 1000 samples from the posterior
posterior_samples(model, n = 1000, lower = 0 upper = 1)
}
