% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.r
\name{prior}
\alias{prior}
\title{Specify a prior}
\usage{
prior(family, ...)
}
\arguments{
\item{family}{the prior distribution (see details)}

\item{...}{see details}
}
\value{
an object of class \code{prior}
}
\description{
Define priors using different different distribution families
}
\details{
\subsection{Available distribution families}{

The following distributions families can be used for the prior
\itemize{
\item \code{normal} a normal distribution
\item \code{uniform} a uniform distribution
\item \code{point} a point
The parameters that need to be specified will be dependent on the
family
}
}

\subsection{normal distribution}{

When \code{family} is set to \code{normal} then the following
parameters must be set
\itemize{
\item \code{mean} mean of the normal prior
\item \code{sd} standard deviation of the normal prior
\item \code{range} (optional) a vector specifying a paramter range
}
}

\subsection{uniform distribution}{

When \code{family} is set to \code{uniform} then the following
parameters must be set
\itemize{
\item \code{min} the lower bound
\item \code{max} the upper bound
}
}

\subsection{point}{

When \code{family} is set to \code{point} then the following
parameters may be set
\itemize{
\item \code{point} the location of the point prior (default: 0)
}
}
}
\examples{
# specify a half-normal (range 0 to Infinity) prior
prior(family = "normal", mean = 0, sd = 13.3, range = c(0, Inf))

# specify a normal prior
prior(family = "normal", mean = 0, sd = 13.3)

# specify a uniform prior
prior(family = "uniform", min = 0, max = 20)

# specify a point prior
prior(family = "point", point = 0)
}
