% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.r
\name{[[,bayesplay-method}
\alias{[[,bayesplay-method}
\title{Get fields from data slot}
\usage{
\S4method{[[}{bayesplay}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{a \code{bayesplay} object}

\item{name}{field name}
}
\value{
content of the named field from the data slot
}
\description{
Get fields from data slot
}
