% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{summary,bf-method}
\alias{summary,bf-method}
\title{Summarise a Bayes factor}
\usage{
\S4method{summary}{bf}(object)
}
\arguments{
\item{object}{a \code{bf} object}
}
\value{
No return, called for side effects
}
\description{
Provide a verbal summary of a Bayes factor and
the level of evidence
}
