% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.r
\name{summary,robustness-method}
\alias{summary,robustness-method}
\title{Summarise a Robustness analysis}
\usage{
\S4method{summary}{robustness}(object)
}
\arguments{
\item{object}{a \code{robustness} object}
}
\value{
No return, called for side effects
}
\description{
Provide a summary of a robustness analysis
}
