% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{visual_compare}
\alias{visual_compare}
\title{Visually compare two models}
\usage{
visual_compare(model1, model2, type = "compare")
}
\arguments{
\item{model1}{a \code{predictive} object}

\item{model2}{a \code{predictive} object}
}
\value{
A \code{ggplot2} object
}
\description{
Visually compare two models
}
\examples{
# define two models
data_model <- likelihood(distribution = "noncentral_d", d, 79)
h0_mod <- prior(distribution = "point", point = 0)
h1_mod <- prior(distribution = "normal", mean = 0, sd = 1)
m0 <- data_model * h0_mod
m1 <- data_model * h1_mod

# visually compare the model
visual_compare(m0, m1, type)
}
