% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.r
\name{bfrr}
\alias{bfrr}
\title{Perform a robustness analysis}
\usage{
bfrr(
  likelihood,
  alternative_prior,
  null_prior,
  parameters,
  steps = 100L,
  cutoff = 3L,
  multicore = TRUE
)
}
\arguments{
\item{likelihood}{\code{likelihood} object representing the data}

\item{alternative_prior}{\code{prior} object for the alternative prior}

\item{null_prior}{\code{prior} object for the null prior}

\item{parameters}{\code{list} of min and max values for each parameter to
vary over.}

\item{steps}{\code{numeric} the number of steps each parameter is varied
over. (Default: 100 steps)}

\item{cutoff}{Minimum Bayes factor value for evidence for the hypothesis
(Default: 3)}

\item{multicore}{Run robustness analysis across multiple cores
(Default: TRUE if available)}
}
\value{
A \code{robustness} object
}
\description{
Perform a robustness analysis by systematically varying the
values of the alternative prior
}
\examples{

# define a likelihood
data_model <- likelihood(family = "normal", mean = 1, sd = 5)

# define the alternative prior
alternative_prior <- prior(family = "normal", mean = 5, sd = 4)

# define the null prior
null_prior <- prior(family = "point", point = 0L)

# set the parameters to vary and the range to vary them across
# vary the mean from -10 to 5
# vary the sd from 1 to 5
parameters <- list(mean = c(-10, 5), sd = c(1, 5))

# mark all Bayes factors larger/smaller than 3/.3 as evidence for the
# alternative / null
cutoff <- 3
bfrr(data_model, alternative_prior, null_prior, parameters, steps = 10,
  cutoff,
  multicore = FALSE
)
}
