% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.r
\name{bfrr}
\alias{bfrr}
\title{Perform a robustness analysis}
\usage{
bfrr(
  likelihood,
  alternative_prior,
  null_prior,
  parameters,
  precision = NULL,
  cutoff = 3L,
  multicore = FALSE
)
}
\arguments{
\item{likelihood}{\code{likelihood} object representing the data}

\item{alternative_prior}{\code{prior} object for the alternative prior}

\item{null_prior}{\code{prior} object for the null prior}

\item{parameters}{\code{list} of min and max values for the parameters to
vary}

\item{precision}{\code{numeric} step size for varied parameter range
(Default: precision is set so that each parameter is varied in 100 steps)}

\item{cutoff}{Minimum Bayes factor value for evidence for the hypothesis}

\item{multicore}{Run robustness analysis across multiple cores}
}
\value{
A \code{robustness} object
}
\description{
Perform a robustness analysis by systematically varying the
values of the alternative prior
}
\examples{

# define a likelihood
data_model <- likelihood(family = "normal", mean = 1, sd = 5)

# define the alternative prior
alternative_prior <- prior(family = "normal", mean = 5, sd = 4)

# define the null prior
null_prior <- prior(family = "point", point = 0L)

# set the parameters to vary and the range to vary them across
# vary the mean from -10 to 5
# vary the sd from 1 to 5
parameters <- list(mean = c(-10, 5), sd = c(1, 5))

# vary the parameters in steps of 0.5
precision <- 0.5

# mark all Bayes factors larger/smaller than 3/.3 as evidence for the
# alternative / null
cutoff <- 3
bfrr(data_model, alternative_prior, null_prior, parameters, precision,
  cutoff,
  multicore = FALSE
)
}
