% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot,prior-method}
\alias{plot,posterior-method}
\alias{plot,likelihood-method}
\alias{plot,product-method}
\alias{plot,prediction-method}
\title{Plot a bayesplay object}
\usage{
\S4method{plot}{bayesplay}(x, y, ...)

\S4method{plot}{prior}(x, y, ...)

\S4method{plot}{posterior}(x, add_prior = FALSE, ...)

\S4method{plot}{likelihood}(x, y, ...)

\S4method{plot}{product}(x, y, ...)

\S4method{plot}{prediction}(x, model_name = "model", ...)
}
\arguments{
\item{x}{a \code{prediction} object}

\item{...}{arguments passed to methods}

\item{add_prior}{set to TRUE to add prior to the posterior plot}

\item{model_name}{name of the model}
}
\value{
a \code{ggplo2} object
}
\description{
Plots an object created by bayesplay
}
