% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot}
\alias{plot}
\alias{plot.prior}
\alias{plot.likelihood}
\alias{plot.posterior}
\alias{plot.product}
\alias{plot.prediction}
\title{Plot a bayesplay object}
\usage{
plot(x, ...)

\method{plot}{prior}(x, ...)

\method{plot}{likelihood}(x, ...)

\method{plot}{posterior}(x, add_prior = FALSE, ...)

\method{plot}{product}(x, ...)

\method{plot}{prediction}(x, ...)
}
\arguments{
\item{x}{a \code{likelihood}, \code{prior}, \code{posterior}, \code{product} or \code{predictive} object #nolint}

\item{...}{arguments passed to methods}

\item{add_prior}{set to TRUE to add prior to the posterior plot}
}
\description{
Plots an object created by bayesplay
}
